refactor: Lotto 객체 위치 원상 복귀
